# public/openapi.yaml
openapi: 3.0.3
info:
  title: Meterra Billing API
  description: |
    Tenant dashboard routes use session cookies (`auth-token`) after `POST /api/auth/login`.
    Admin routes use `admin_token` after `POST /api/admin/auth/login`.

    **Usage ingestion** uses the `X-API-Key` header (and optional project on the key or `projectId` in the body).

    Download this spec: [`/openapi.yaml`](/openapi.yaml)
  version: 1.0.0
servers:
  - url: /
    description: Current host (same origin)
...
security:
  - TenantSession: []
paths:
...
  /api/pricing:
    get:
      tags: [Pricing]
      summary: List pricing plans with rules
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: "#/components/schemas/PricingPlan"
    post:
      tags: [Pricing]
      summary: Create pricing plan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingPlanCreate"
      responses:
        "200":
          description: Created
...
  /api/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
    post:
      tags: [Subscriptions]
      summary: Create subscription
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [customerId, pricingPlanId, startDate]
              properties:
                customerId:
                  type: string
                  format: uuid
                pricingPlanId:
                  type: string
                  format: uuid
                billingCycle:
                  type: string
                  enum: [MONTHLY, YEARLY, QUARTERLY, PAY_AS_YOU_GO]
                startDate:
                  description: ISO date or datetime (e.g. from HTML date input)
                  type: string
                  example: "2026-03-29"
                endDate:
                  type: string
                  format: date-time
                metadata:
                  type: object
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    $ref: "#/components/schemas/Subscription"
        "404":
          description: Customer or plan not found
  /api/usage/events:
    post:
      tags: [Usage]
      summary: Record usage event
      description: |
        Requires `X-API-Key`. Project is taken from the key unless `projectId` is sent in the body (tenant-wide keys).
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageEventCreate"
      responses:
        "200":
          description: Event stored and rated
          content:
            application/json:
              schema:
                type: object
                required: [success, eventId, estimatedCost, pricingRuleId]
                properties:
                  success:
                    type: boolean
                  eventId:
                    type: string
                    format: uuid
                  estimatedCost:
                    type: number
                  pricingRuleId:
                    type: string
                    format: uuid
                    nullable: true
        "400":
          description: Validation or missing projectId
        "401":
          description: Invalid API key
        "404":
          description: Project or product not found
...
components:
  securitySchemes:
    TenantSession:
      type: apiKey
      in: cookie
      name: auth-token
    AdminSession:
      type: apiKey
      in: cookie
      name: admin_token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    PricingPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        rules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRule"
    PricingPlanCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          type: string
          format: date-time
        rules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleCreate"
    PricingRule:
      type: object
      properties:
        id:
          type: string
        eventName:
          type: string
        metric:
          type: string
        pricingType:
          type: string
          enum: [FLAT, PER_UNIT, TIERED, VOLUME]
        unitPrice:
          type: string
          nullable: true
        flatRate:
          type: string
          nullable: true
        productId:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
    PricingRuleCreate:
      type: object
      required: [eventName, metric, pricingType]
      properties:
        eventName:
          type: string
        metric:
          type: string
        pricingType:
          type: string
          enum: [FLAT, PER_UNIT, TIERED, VOLUME]
        flatRate:
          type: number
        unitPrice:
          type: number
        tiers:
          type: array
          items:
            type: object
            required: [upTo, price]
            properties:
              upTo:
                type: number
              price:
                type: number
        minQuantity:
          type: number
        maxQuantity:
          type: number
        productId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        pricingPlanId:
          type: string
        status:
          type: string
        billingCycle:
          type: string
        startDate:
          type: string
          format: date-time
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
    UsageEventCreate:
      type: object
      required: [eventName, quantity, unit, timestamp]
      properties:
        eventName:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
        timestamp:
          type: string
          format: date-time
        properties:
          type: object
          additionalProperties: true
        idempotencyKey:
          type: string
        projectId:
          type: string
          format: uuid
          description: Required when API key has no project scope
        productId:
          type: string
          format: uuid
          description: Optional; matches product-scoped pricing rules
...